#2226 Move Task types file to core/types & import update
