+++ core/lib/sync/task.ts
@@ -7,7 +7,7 @@ import fetch from 'isomorphic-fetch'
 import * as localForage from 'localforage'
 import rootStore from '@vue-storefront/store'
 import { adjustMultistoreApiUrl, currentStoreView } from '@vue-storefront/store/lib/multistore'
-import Task from '@vue-storefront/store/types/task/Task'
+import Task from '@vue-storefront/core/lib/sync/types/Task'
 import { Logger } from '@vue-storefront/core/lib/logger'
 import { TaskQueue } from '@vue-storefront/core/lib/sync'
 import * as entities from '@vue-storefront/store/lib/entities'
